requirements.txt
streamlit
pandas
numpy
kafka-python
psycopg2-binary
sqlalchemy
